[WIP] organize helper functions into utils
