+++ src/canvas/canvasUtils.js
@@ -18,7 +18,7 @@ import usePanAndZoom, {
   durationToWidth,
   positionToTimestamp,
   timestampToPosition,
-} from '../usePanAndZoom';
+} from '../utils/usePanAndZoom';
 
 // hidpi canvas: https://www.html5rocks.com/en/tutorials/canvas/hidpi/
 export function configureRetinaCanvas(
