+++ src/utils/preprocessData.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { TimelineEvent } from './speedscope/import/chrome';
+import type { TimelineEvent } from '../speedscope/import/chrome';
 
 import type {
   Milliseconds,
