+++ src/utils/preprocessFlamechart.js
@@ -1,10 +1,10 @@
 // @flow
 
-import { importFromChromeTimeline } from './speedscope/import/chrome';
-import { Flamechart } from './speedscope/lib/flamechart';
+import { importFromChromeTimeline } from '../speedscope/import/chrome';
+import { Flamechart } from '../speedscope/lib/flamechart';
 
-import type { TimelineEvent } from './speedscope/import/chrome';
-import type { FlamechartData } from './types';
+import type { TimelineEvent } from '../speedscope/import/chrome';
+import type { FlamechartData } from '../types';
 
 export default function preprocessFlamechart(
   rawData: TimelineEvent[]
